feat(app-service-configurable): Create an Application Service which create its Pipeline from Configuration
Need a way to define/configure a function pipeline via configuration without requiring a developer/user to compile their own application. This configuration can be specified in TOML with support for live changes from a registry (i.e. Cobe able to dynamically adjust filters and functions to execute as EdgeX data is received. Profiles can be used to deploy multiple instances of the application service with different function pipelines

Closes #108

Signed-off-by: Lenny Goodell <leonard.goodell@inyel.com>