fix(snap): allow disabling of secret-store via config hook
This fix adds a new snap configure option 'security-secret-store'
that controls how secrets are stored. By default if a profile's
configuration file requires secrets, secret-store will be used.
If 'security-secret-store' is set to 'false', then insecure
secrets will be used instead. This aligns the snap with the same
capabilty available in the main edgexfoundry snap and the standard
docker-compose files.

Fixes: 150

Signed-off-by: Tony Espy <espy@canonical.com>